voice_qa/requirements.txt
openai-whisper
pyttsx3
numpy
requests
openai
python-dotenv